Armor System
Added an armor system

Armor system

Everything that can take damage has an innate "armor" value

Every bullet has an innate armor piercing or "ap" value

When something takes damage, it subtracts its "armor" value from the bullets "ap" value.

When a bullet has 0 "ap" value left, it is destroyed on contact.

Bullets with any amount of remaining "ap" value will continue through a normal object that can take damage.

"ap" value is determined by the weapon shooting it. When a bullet is pulled from the object pool, its "ap" value is given to it by the gun. When a bullet returns to the object pool, its "ap" value is set back to a default of 0.

Armor does not reduce the amount of damage an enemy takes.

HealthComponent:
TakeDamage() should be updated to the following:
accept an armor piercing value
Get the value of the health component, and subtract it from the bullets armor piercing value, then update that value on the bullet.
Example:
A bullet with AP of 50 collides with an enemy with an armor value of 10.
The bullet has its AP reduced to 40.
The bullet will continue on its trajectory.

Fixed shotgun damage.